rename docker->docker_utils to avoid import/aliasing issues
